get rid of handover_other, windows supports AF_UNIX
