+++ lib/resumption/handover_test.go
@@ -1,5 +1,3 @@
-//go:build unix
-
 // Teleport
 // Copyright (C) 2024 Gravitational, Inc.
 //
